test(e2e): revert new specs to fit nightly under TC 20-min cap
