+++ e2e/README.md
@@ -26,14 +26,6 @@ site `https://e2e.newspackstaging.com`. The build definition lives in TeamCity
 settings, not in this repo; its steps are: install dependencies, write a `.env`,
 then run `npm run test:setup` (the setup projects provision the site over SSH).
 
-That staging site is pinned to the **stable release** channel, and provisioning
-rebuilds against the plugin version installed there – not the version the specs
-were merged with. Write specs against the release-channel UI: a spec that drives
-UI which only exists in `alpha`/`main` will fail nightly until that feature ships
-to stable. Don't work around that with channel branching or skips – hold coverage
-for alpha-only UI until it reaches release. See `AGENTS.md` → "Site setup model"
-for details.
-
 [Credentials for the Atomic site used for the e2e testing.](https://mc.a8c.com/secret-store/?secret_id=12168)
 
 The build is parameterised by these variables (set in TeamCity, not committed):
@@ -120,16 +112,6 @@ Tests can be written by hand in the `tests` directory, or with the help of Playw
 Tag tests `@vanilla` or `@with-woo` so they run in the matching phase. If a test
 needs a particular fixture (a page, product, user, …), have `e2e-setup.sh` (or the
 underlying `site-setup.sh`) create it, so it's rebuilt on every run.
-
-### Content gating (Access Control)
-
-Provisioning defines `NEWSPACK_CONTENT_GATES` in wp-config, enabling the
-Audience > Access control wizard and its front-end enforcement, which
-`content-gating.spec.ts` and `premium-newsletters.spec.ts` cover. In the
-`--woo` phase, `e2e-setup.sh` deactivates `woocommerce-memberships` after the
-generic bootstrap: Access Control defers to Memberships whenever that plugin
-is active, so leaving it on would make the first-party gating inert (this also
-matches the target state of migrated Newspack sites).
 
 ## Provisioning the test site manually
